requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
dj-database-url==2.1.0
Pillow==10.4.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
requests==2.31.0
google-auth==2.26.1
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.5.0
django-storages[google]==1.13.2
boto3==1.34.0
django-admin-interface==0.19.0
django-ratelimit==4.1.0
django-axes==5.26.0
django-silk==5.0.1
django-debug-toolbar==4.1.0
django-extensions==3.2.1
werkzeug==3.1.0
django-redis==5.4.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.8.0
django-celery-results==2.5.0
django-guardian>=3.2.0
django-model-utils>=4.5.0
django-taggit>=5.0.0
django-taggit-serializer>=0.1.7
django-templated-mail>=0.1.0
djoser==2.3.0
drf-yasg>=1.21.4
drf-spectacular>=0.27.4
drf-spectacular-sidecar>=0.21.0
django-restql>=0.14.0
django-allauth>=0.63.0
django-oauth-toolkit>=3.0.0
social-auth-app-django>=5.4.0
social-auth-core>=4.5.0
razorpay==1.3.0